certification3/requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.5
seaborn>=0.11
scikit-learn>=1.3
SQLAlchemy>=2.0
scipy>=1.10
psycopg2-binary>=2.9
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0